First attempt to pull data.
This really just pulls down name and dps of the top 100.  The page returns if there are more pages as true, so I'll need to interate across pages.  This will be a lot of data, so I'll get the formatting down for a low number and progressively increase until either the PC crashes/stalls or I have all the data.  It shouldn't be too big though.  I'd assume I just need the DPS, Class, Spec, Race, and ilvl for each boss.  That's 5 cells per boss over at least 100,000 characters.  I'll also want to do more such as pull from Zones to create a list of fight ID's so that this can be raid agnostic.